fix(portability): 修复 Windows/MSVC 编译失败（min/max 宏、窄整型分布、进程 id）
Windows 构建（MSVC、/W4）在三处中止：
- socket_platform.h：winsock2.h 间接引入的 windows.h 定义 min/max 宏，把 dns.cpp
  的 std::max/std::min 展开成语法错误（C2589）。包含 Winsock 前定义 NOMINMAX。
- websocket.cpp：std::uniform_int_distribution<uint8_t> 不是合法实例化
  （[rand.req.genl] 排除 char/unsigned char/int8_t/uint8_t，MSVC C2338）；
  改为按 unsigned int 抽样后显式收窄（客户端 key 与掩码 key 两处）。
- 测试里的 ::getpid() 在 MSVC 上需要 <process.h>/_getpid，<unistd.h> 不存在
  （idb_test.cpp C1083）。6 个测试文件统一为 CurrentProcessId() 助手；
  未受保护的 <unistd.h> 移入平台分支（dom_binding_test.cpp、ui_smoke_test.cpp）。

验证（GCC 本机）：neko_network_tests 73/73、neko_storage_tests 50/50、
neko_ui_tests 23/23、neko_javascript_tests 274/274（排除 DomBinderXhrTest，
该套件在本机既有中止，git stash 基线对比一致）。

+++ src/network/src/websocket.cpp
@@ -57,10 +57,13 @@ std::string GenerateWebSocketKey()
 {
   std::random_device rd;
   std::mt19937 gen(rd());
-  std::uniform_int_distribution<uint8_t> dist(0, 255);
+  // A distribution over uint8_t is not conforming (char / unsigned char /
+  // int8_t / uint8_t are excluded by [rand.req.genl]); draw in unsigned int
+  // and narrow explicitly.
+  std::uniform_int_distribution<unsigned int> dist(0, 255);
   std::array<uint8_t, 16> bytes = {};
   for (auto& byte : bytes) {
-    byte = dist(gen);
+    byte = static_cast<uint8_t>(dist(gen));
   }
   return Base64Encode(std::string_view(reinterpret_cast<const char*>(bytes.data()), bytes.size()));
 }
@@ -405,9 +408,10 @@ base::Status WebSocket::SendFrame(uint8_t opcode, std::string_view payload, bool
   if (mask) {
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_int_distribution<uint8_t> dist(0, 255);
+    // See GenerateWebSocketKey(): the distribution draws in unsigned int.
+    std::uniform_int_distribution<unsigned int> dist(0, 255);
     for (auto& b : masking_key) {
-      b = dist(gen);
+      b = static_cast<uint8_t>(dist(gen));
       frame.push_back(static_cast<char>(b));
     }
   }
